introduce os::Linux::dll_path, add dlerror output, add some quoting
