+++ src/hotspot/os/posix/os_posix.cpp
@@ -54,9 +54,6 @@
 #include <dlfcn.h>
 #include <grp.h>
 #include <locale.h>
-#ifdef LINUX
-#include <link.h>
-#endif
 #include <netdb.h>
 #include <pwd.h>
 #include <pthread.h>
@@ -709,25 +706,24 @@ void* os::dll_lookup(void* handle, const char* name) {
 }
 
 void os::dll_unload(void *lib) {
-#if defined(LINUX)
-  struct link_map *lmap;
-  int res_dli = ::dlinfo(lib, RTLD_DI_LINKMAP, &lmap);
-  const char* l_path = "<none>";
-  if (res_dli == 0) {
-    l_path = lmap->l_name;
-  }
-#else
-  const char* l_path = "<not available>";
-#endif
-
+  const char* l_path = LINUX_ONLY(os::Linux::dll_path(lib))
+                       NOT_LINUX("<not available>");
   int res = ::dlclose(lib);
 
   if (res == 0) {
-    Events::log_dll_message(NULL, "Unloaded shared library %s [" INTPTR_FORMAT "]", l_path, p2i(lib));
+    Events::log_dll_message(NULL, "Unloaded shared library \"%s\" [" INTPTR_FORMAT "]",
+                            l_path, p2i(lib));
     log_info(os)("Unloaded shared library \"%s\" [" INTPTR_FORMAT "]", l_path, p2i(lib));
   } else {
-    Events::log_dll_message(NULL, "Attempt to unload shared library %s [" INTPTR_FORMAT "] failed", l_path, p2i(lib));
-    log_info(os)("Attempt to unload shared library \"%s\" [" INTPTR_FORMAT "] failed", l_path, p2i(lib));
+    const char* error_report = ::dlerror();
+    if (error_report == NULL) {
+      error_report = "dlerror returned no error description";
+    }
+
+    Events::log_dll_message(NULL, "Attempt to unload shared library \"%s\" [" INTPTR_FORMAT "] failed, %s",
+                            l_path, p2i(lib), error_report);
+    log_info(os)("Attempt to unload shared library \"%s\" [" INTPTR_FORMAT "] failed, %s",
+                  l_path, p2i(lib), error_report);
   }
 }
